+++ test/units/phonenumber_test.dart
@@ -13,5 +13,22 @@ void main() {
       expect(PhoneNumber.parse('994(55)566-12-33').isValid, true);
       expect(PhoneNumber.parse('9945566').isValid, false);
     });
+
+    test('.country', () {
+      expect(PhoneNumber.parse('+99455777').country.code, 'AZ');
+      expect(PhoneNumber.parse('+4345555').country.code, 'AT');
+    });
+
+    test('.nationalNumber', () {
+      expect(PhoneNumber.parse('+38612345678').nationalNumber, '12345678');
+      expect(PhoneNumber.parse('+13456789').nationalNumber, '3456789');
+    });
+
+    test('.formattedNumber', () {
+      expect(
+        PhoneNumber.parse('994 11-2345676').formattedNumber,
+        '+994112345676',
+      );
+    });
   });
 }